% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{create_key}
\alias{create_key}
\title{Create Key}
\usage{
create_key(keyname)
}
\arguments{
\item{keyname}{The key name is for debugging purposes and is not required to
be unique.}
}
\value{
A unique string representing the newly created key.
}
\description{
To allow cross-cutting concern to control access to their local state, the
RuntimeContext API provides a function which takes a keyname as input, and
returns a unique key.
}
