% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_current}
\alias{get_current}
\title{Get Current Context}
\usage{
get_current()
}
\value{
The current \code{Context} object.
}
\description{
To access the context associated with program execution, the Context API
provides a function which takes no arguments and returns a Context.#'
}
