% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_propagation.R
\name{get_current_span}
\alias{get_current_span}
\title{Get the current span}
\usage{
get_current_span(span, context = NULL)

get_current_span(span, context = NULL)
}
\arguments{
\item{span}{The Span to set.}

\item{context}{The context in which to set the current span.}
}
\value{
The current span in the context or \code{invalid_span} if no span is
found.

The updated context
}
\description{
Get the current span

Set the current span
}
