% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_value}
\alias{get_value}
\title{Get Value}
\usage{
get_value(key, context = NULL)
}
\arguments{
\item{keyname}{The key name is for debugging purposes and is not required to
be unique.}
}
\value{
A unique string representing the newly created key.
}
\description{
To access the local state of a concern, the RuntimeContext API provides a
function which takes a context and a key as input. and returns a value.
}
