% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentelemetry-package.R
\docType{package}
\name{opentelemetry-package}
\alias{opentelemetry}
\alias{opentelemetry-package}
\title{opentelemetry: OpenTelemtry R API and SDK}
\description{
Implementation of OpenTelemetry in R.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{Alex@AlexAxthelm.com} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

}
\keyword{internal}
