% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resource.R
\name{Resource}
\alias{Resource}
\title{R6 Class Representing a Resource}
\description{
A Resource is an immutable representation of the entity producing telemetry
as Attributes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Resource-new}{\code{Resource$new()}}
\item \href{#method-Resource-create}{\code{Resource$create()}}
\item \href{#method-Resource-get_empty}{\code{Resource$get_empty()}}
\item \href{#method-Resource-merge}{\code{Resource$merge()}}
\item \href{#method-Resource-to_json}{\code{Resource$to_json()}}
\item \href{#method-Resource-clone}{\code{Resource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resource-new"></a>}}
\if{latex}{\out{\hypertarget{method-Resource-new}{}}}
\subsection{Method \code{new()}}{
Create a new resource object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$new(attributes = list(), schema_url = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attributes}}{Resource Attributes.}

\item{\code{schema_url}}{URL for schema describing Resource attributes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Resource} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resource-create"></a>}}
\if{latex}{\out{\hypertarget{method-Resource-create}{}}}
\subsection{Method \code{create()}}{
Create a new resource object (alias for \code{Resource$new()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$create(attributes = list(), schema_url = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attributes}}{Resource Attributes.}

\item{\code{schema_url}}{URL for schema describing Resource attributes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Resource} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resource-get_empty"></a>}}
\if{latex}{\out{\hypertarget{method-Resource-get_empty}{}}}
\subsection{Method \code{get_empty()}}{
TODO: Document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resource-merge"></a>}}
\if{latex}{\out{\hypertarget{method-Resource-merge}{}}}
\subsection{Method \code{merge()}}{
TODO: Document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$merge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resource-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-Resource-to_json}{}}}
\subsection{Method \code{to_json()}}{
TODO: Document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Resource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
