% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{set_value}
\alias{set_value}
\title{Set Value}
\usage{
set_value(key, value, context = NULL)
}
\arguments{
\item{key}{The key name is for debugging purposes and is not required to be
unique.}

\item{value}{The value to set.}

\item{context}{The context in which to set the value. If not provided, the
current context is used.}
}
\value{
A unique string representing the newly created key.
}
\description{
To record the local state of a cross-cutting concern, the RuntimeContext API
provides a function which takes a context, a key, and a value as input, and
returns an updated context which contains the new value.'
}
